mandir-photo-ai/face-backend/requirements.txt
flask
keras_facenet
faiss-cpu
numpy
pillow
